Ver 3.5.17
[Ingame]
- 대출 팝업 버그 수정
[Play]
- 일러스트 팝업 이펙트 안나오던 버그 수정

+++ Stage_OnAir/Assets/02_Ingame/Scripts/Manager/GameManager.cs
@@ -331,7 +331,7 @@ public class GameManager : Singleton<GameManager>
             }
             if (data.Keys.Contains("BestQuality"))
             {
-                Best_Quality = int.Parse(data["BestQuality"]["N"].ToString());
+                Best_Quality = float.Parse(data["BestQuality"]["N"].ToString());
             }
             if (data.Keys.Contains("BestAudience"))
             {
